refactor: 迁移 AssetStudio 导入 Worker
