+++ xl_updata_tool/docs/代码所有权与边界.md
@@ -10,7 +10,7 @@
 | 角色 | `app/features/characters/` | `page.py`、`controller.py`、`service.py`；兼容入口为 `app/core/character_*`、`app/ui/views/character_view.py` | Lua 解析、角色仓库、Wiki 页面、未读和 CSV；不直接改音频或版本 |
 | 预览 | `app/features/preview/` | `page.py`、`controller.py`、`service.py`、`item.py`、`worker.py`；兼容入口为 `app/core/preview_catalog.py`、`app/ui/features/fgui_atlas.py`、Spine 适配器 | 图片、FGUI、Spine、视频导出 |
 | 版本 | `app/features/versions/` | `page.py`、`controller.py`、`service.py`、`worker.py`；兼容入口为 `app/core/version_*`、`app/ui/workers/download.py` | 版本列表、检查更新、下载、删除和 Bundle 状态 |
-| 导入 | `app/features/importer/` | `spec.py`、`service.py`、`controller.py`、`postprocessing.py`、`worker.py`；兼容执行器为 `app/ui/workers/import_as.py` | Bundle、AssetStudio、导出计划、发布事务和后处理结果 |
+| 导入 | `app/features/importer/` | `spec.py`、`service.py`、`controller.py`、`postprocessing.py`、`processing.py`、`worker.py`；兼容入口为 `app/ui/workers/import_as.py` | Bundle、AssetStudio、导出计划、发布事务和后处理结果 |
 | 平台与诊断 | `app/platform/` | `app/core/path_utils.py`、`logger.py`、`database.py` 等 | 路径、文件、数据库、进程、网络、日志和诊断 |
 | 共享 UI | `app/shared/qt/` | `app/ui/theme.py`、`widgets/view_chrome.py` | 主题令牌、页头、命令栏、空状态和通用控件 |
 | Shell 与装配 | `app/shell/`、`app/bootstrap/` | `app/ui/main_window.py`、`main.py` | 导航、任务宿主、Feature 注册和运行时装配 |
